@@ -5,20 +5,22 @@
    "devDependencies": {
       "@types/node": "^8.0.29",
       "ts-node": "^9.1.1",
-      "typescript": "3.3.3333"
+      "typescript": "^3.9.9"
    },
    "dependencies": {
       "@types/express": "^4.17.11",
       "body-parser": "^1.19.0",
+      "class-validator": "^0.13.1",
       "express": "^4.17.1",
       "morgan": "^1.10.0",
       "mysql": "^2.18.1",
       "nodemon": "^2.0.7",
       "reflect-metadata": "^0.1.13",
-      "typeorm": "^0.2.32"
+      "typeorm": "^0.2.32",
+      "uuid": "^8.3.2"
    },
    "scripts": {
-      "start": "ts-node src/index.ts",
+      "start": "3-node src/index.ts",
       "typeorm": "ts-node ./node_modules/typeorm/cli.js"
    }
 }
