@@ -1,50 +1,120 @@
-import {getConnection,getRepository,getConnectionManager,createConnection} from "typeorm";
-import {Artista} from "../../entity/Artista";
-
+import {getConnection,getRepository,getConnectionManager,createConnection, Like} from "typeorm";
+import {Artista} from "../entity/Artista";
+import {Album} from "../entity/Album";
+import {v4 as uuidv4} from "uuid";
+import {validateOrReject} from "class-validator";
+import {MensajesManager} from "../../Utilities/MensajesManager/MensajesManager";
 
 export class ArtistasRepository {
 
-    public  async  crearArtista(datosartista){
-  
-         
+    public  async  crearArtista(datosartista:Artista):Promise<any>{
+          
         try{
-            await createConnection();
             
-           let  artista = new Artista()
-            artista.nombre = datosartista.nombre;
+            await createConnection();    
+            let  artista = new Artista()
+            artista.id = uuidv4();
+	        artista.nombre = datosartista.nombre;
             artista.nombreArtistico = datosartista.nombreArtistico;
-            artista.fechaDeNacimiento = datosartista.fechaDeNacimiento;
+            artista.anoDeNacimiento = datosartista.anoDeNacimiento;
             artista.nacionalidad = datosartista.nacionalidad;
             artista.web = datosartista.web;
+            artista.fkIdEstatus = 1;
+            try{
+                await validateOrReject(artista);
+            }catch(excepcionDeValidacion){
+                
+                return MensajesManager.crearMensajeDeErrorDeValidacion(excepcionDeValidacion);
+            }finally{
+                
+            }
             const user =await getConnection().manager.save(artista);
-         
-            console.log("artista guardado exitosamente: "+artista.nombre);
+            
+            
         }catch(excepcion){
-            console.log(excepcion);
+           
+           return MensajesManager.crearMensajeDeError(excepcion);
+        }finally{
+            getConnection().close();
         } 
+        return MensajesManager.crearMensajeDeExito("artista registrado con exito");
     }
 
-    public async actualizarArtista (artistaP:Artista):Promise<Artista>{
-            
+    public async actualizarArtista (artistaP:Artista):Promise<any>{
         try{
             await createConnection();
             const artista =await getRepository(Artista).findOne(artistaP.id);
             if(artista == null){
-                return null;
+                return MensajesManager.crearMensajeDeErrorDeValidacion(null);
             }
             artista.nombre = artistaP.nombre;
             artista.nombreArtistico = artistaP.nombreArtistico;
-            artista.fechaDeNacimiento = artistaP.fechaDeNacimiento;
-            artista.nacionalidad = artistaP.nacionalidad;
+            artista.anoDeNacimiento = artistaP.anoDeNacimiento;
             artista.web = artistaP.web;
-            artista.estado = artistaP.estado;
+            artista.nacionalidad = artistaP.nacionalidad;
+            artista.fkIdEstatus = artistaP.fkIdEstatus;
+            try{
+                await validateOrReject(artista);
+            }catch(excepcionesDeValidacion){ 
+                return MensajesManager.crearMensajeDeErrorDeValidacion(excepcionesDeValidacion);
+            }
             await getRepository(Artista).save(artistaP);
-            console.log("artista actualizado exitosamente: "+artistaP.nombre);
         }catch(excepcion){
-            console.log(excepcion);
+            return MensajesManager.crearMensajeDeErrorDeValidacion(excepcion);
+        }finally{
+            getConnection().close();
         } 
-
+        return MensajesManager.crearMensajeDeExito("datos modificados con exito")
 
     }
+
+    public async buscarArtistaPorId(idArtista:string):Promise<any>{
+        
+        
+        let artista;
+        try{
+            await createConnection();
+            artista = await getRepository(Artista).findOneOrFail({where:{id:idArtista}});
+            if(artista == (null|| undefined)){
+                return MensajesManager.crearMensajeDeErrorDeValidacion(null);
+            }            
+        }catch(excepcion){
+            
+            return MensajesManager.crearMensajeDeError(excepcion);
+        
+        }finally{
+            getConnection().close();
+        }
+    
+        return MensajesManager.crearMensajeDeExito("consulta exitosa",artista);
+
+    }
+
+    public async buscarArtistaPorNombre(nombreArtista:string):Promise<any>{
+        
+       
+        let artistas = null;
+        let ar = [];
+    
+        try{
+            await createConnection();
+            artistas = await getRepository(Artista).find({nombreArtistico:Like("%"+nombreArtista+"%")});
+            if(artistas == null || artistas.length == 0){
+                return MensajesManager.crearMensajeDeErrorDeValidacion(null);
+            }
+        }catch(excepcion){
+            getConnection().close();
+            return MensajesManager.crearMensajeDeErrorDeValidacion(excepcion);
+        }finally{
+            getConnection().close();
+        }
+       
+       
+        return MensajesManager.crearMensajeDeExito("consulta realizada con exito",artistas);
+
+    }
+
+
+   
 }
 
